rename block_group page as climate. Kiera text
