+++ src/app/address/english.js
@@ -18,8 +18,8 @@ export const english = {
   map_view: 'Map view',
   neighborhood: 'Neighborhood',
   no_neighborhood_name: 'No neighborhood name',
-  block_group: ' Climate Threats and Vulnerability',
-  no_block_group: 'Unknown Census Block Group',
+  climate: ' Climate Threats and Vulnerability',
+  no_climate: 'Unknown Census Block Group',
   no_city_pickup: 'No city pickup',
   no_information_available: 'No information available',
   no_results_found: 'No results found',
