+++ src/app/address/spanish.js
@@ -18,8 +18,8 @@ export const spanish = {
   map_view: 'Mapa',
   neighborhood: 'Barrio',
   no_neighborhood_name: 'Barrio sin nombre',
-  block_group: 'Censo Block Group',
-  no_block_group: 'Censo Block Group sin nombre',
+  climate: 'Censo Block Group',
+  no_climate: 'Censo Block Group sin nombre',
   no_city_pickup: 'no proporcionada por la ciudad de Asheville',
   no_information_available: 'No hay informaci\xF3n disponisble',
   no_results_found: 'No se ha encontrado resultados',
