+++ src/app/climate/ClickableTile.js
